Update timesheets_app_project_patch.rb to fix #73
In Redmine 3.2 the shared_versions method (in project.rb) is now called when a new project is created from a new method in (projects_helper.rb).  In the shared_versions method there is a conditional:

def shared_versions
    if new_record?
        Version.
      .......
    else
      @shared_versions ||= begin
        .........
      end
    end
end

Because new_record? is true, @shared_versions is never defined which then passed a nil to shared_versions_with_timelogs.

The fix is to ensure that @shared_versions is always defined.